+++ config/script_test.go
@@ -742,7 +742,7 @@ func testScriptEval(t *testing.T, tcase scriptTestcase) {
 
 	st, err := rootTree.Stack()
 	assert.NoError(t, err)
-	hclctx := eval.NewContext(stdlib.Functions(tempdir))
+	hclctx := eval.NewContext(stdlib.Functions(tempdir, []string{}))
 	hclctx.SetNamespace("global", tcase.globals)
 	runtime := cfg.Runtime()
 	runtime.Merge(st.RuntimeValues(cfg))